# libs/py/contracts-generated/src/cvm_contracts_generated/platform_api.py
# generated by datamodel-codegen:
#   filename:  platform-api.openapi.yaml
#   timestamp: 2026-03-24T13:58:11+00:00

from __future__ import annotations

from enum import StrEnum
from typing import Any

from pydantic import AwareDatetime, BaseModel, ConfigDict


class CreateCaseRequest(BaseModel):
    title: str
    ownerTeamId: str


class CreateCaseResponse(BaseModel):
    caseId: str
    status: str


class CreateJdVersionRequest(BaseModel):
    rawText: str
    source: str


class CreateJdVersionResponse(BaseModel):
    jdVersionId: str
    versionNo: int
    status: str


class EvidenceRef(BaseModel):
    label: str
    excerpt: str


class StructuredFilters(BaseModel):
    model_config = ConfigDict(
        extra='allow',
    )
    location: str | list[str] | None = None
    degree: int | None = None
    schoolType: int | None = None
    workExperienceRange: int | None = None
    page: int | None = None
    pageSize: int | None = None


class ConditionPlanDraft(BaseModel):
    mustTerms: list[str]
    shouldTerms: list[str]
    excludeTerms: list[str]
    structuredFilters: StructuredFilters
    evidenceRefs: list[EvidenceRef]


class CreateKeywordDraftJobRequest(BaseModel):
    jdVersionId: str
    modelVersion: str
    promptVersion: str


class CreateKeywordDraftJobResponse(BaseModel):
    jobId: str
    status: str
    planId: str
    draft: ConditionPlanDraft


class ConfirmConditionPlanRequest(ConditionPlanDraft):
    confirmedBy: str


class ConfirmConditionPlanResponse(BaseModel):
    planId: str
    status: str
    normalizedQuery: dict[str, Any]


class CreateSearchRunRequest(BaseModel):
    caseId: str
    planId: str
    pageBudget: int
    idempotencyKey: str


class CreateSearchRunResponse(BaseModel):
    runId: str
    status: str


class SearchRunStatusResponse(BaseModel):
    runId: str
    status: str
    progress: float
    pageCount: int
    errorSummary: str | None = None


class CandidateCard(BaseModel):
    candidateId: str
    externalIdentityId: str
    name: str
    title: str
    company: str
    location: str
    summary: str | None = None


class SearchRunPageSnapshot(BaseModel):
    pageNo: int
    status: str
    fetchedAt: AwareDatetime
    candidates: list[CandidateCard]
    total: int
    errorCode: str | None = None
    errorMessage: str | None = None


class SearchRunPagesResponse(BaseModel):
    runId: str
    snapshots: list[SearchRunPageSnapshot]


class ResumeSnapshot(BaseModel):
    id: str
    content: dict[str, Any]


class ResumeAnalysis(BaseModel):
    status: str
    summary: str
    evidenceSpans: list[str]
    riskFlags: list[str]


class Verdict(StrEnum):
    Match = 'Match'
    Maybe = 'Maybe'
    No = 'No'


class SaveVerdictRequest(BaseModel):
    verdict: Verdict
    reasons: list[str]
    notes: str | None = None
    actorId: str
    resumeSnapshotId: str | None = None


class SaveVerdictResponse(BaseModel):
    candidateId: str
    latestVerdict: Verdict
    updatedAt: AwareDatetime


class MaskPolicy(StrEnum):
    masked = 'masked'
    sensitive = 'sensitive'


class CreateExportRequest(BaseModel):
    caseId: str
    maskPolicy: MaskPolicy
    reason: str
    idempotencyKey: str


class CreateExportResponse(BaseModel):
    exportJobId: str
    status: str
    filePath: str | None = None


class MetricItem(BaseModel):
    name: str
    value: float


class OpsVersionInfo(BaseModel):
    api: str
    apiBuildId: str
    workerBuildId: str
    externalCts: str
    temporalNamespace: str
    temporalUiUrl: str
    temporalVisibilityBackend: str


class OpsSummaryResponse(BaseModel):
    queue: dict[str, Any]
    failures: dict[str, Any]
    latency: dict[str, Any]
    version: OpsVersionInfo
    metrics: list[MetricItem]


class TemporalSearchRunDiagnosticResponse(BaseModel):
    runId: str
    workflowId: str
    namespace: str
    taskQueue: str
    appStatus: str
    temporalExecutionFound: bool
    temporalExecutionStatus: str | None = None
    visibilityIndexed: bool
    visibilityBackend: str
    startedAt: AwareDatetime | None = None
    closedAt: AwareDatetime | None = None
    error: str | None = None
    temporalUiUrl: str


class CreateEvalRunRequest(BaseModel):
    suiteId: str
    datasetId: str
    targetVersion: str


class CreateEvalRunResponse(BaseModel):
    evalRunId: str
    status: str


class VerdictRecord(BaseModel):
    verdict: Verdict
    reasons: list[str]
    notes: str | None = None
    actorId: str
    createdAt: AwareDatetime


class CandidateDetailResponse(BaseModel):
    candidate: CandidateCard
    resumeSnapshot: ResumeSnapshot
    aiAnalysis: ResumeAnalysis
    verdictHistory: list[VerdictRecord]