# libs/ts/api-client-generated/src/generated/models/OpsVersionInfo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OpsVersionInfo = {
    api: string;
    apiBuildId: string;
    workerBuildId: string;
    externalCts: string;
    temporalNamespace: string;
    temporalUiUrl: string;
    temporalVisibilityBackend: string;
};